+++ packages/router/src/lib/helpers/metrics.ts
@@ -1,6 +1,5 @@
 import { createLoggingContext, jsonifyError, RequestContext } from "@connext/nxtp-utils";
 import { constants, BigNumber, utils } from "ethers";
-import { getDeployedPriceOracleContract } from "../../config";
 
 import { getContext } from "../../router";
 import {
@@ -32,10 +31,6 @@ export const convertToUsd = async (
   requestContext: RequestContext,
 ): Promise<number> => {
   const { txService, logger } = getContext();
-
-  // Get token price
-  const priceOracleContract = getDeployedPriceOracleContract(chainId);
-  if (!priceOracleContract || !priceOracleContract.address) return 0;
 
   const price = await txService.getTokenPrice(chainId, assetId, undefined, requestContext);
   if (price.isZero()) {