+++ packages/router/test/globalTestHook.ts
@@ -12,10 +12,14 @@ import {
   chainDataMock,
 } from "./utils";
 import { Context } from "../src/router";
+import { Web3Signer } from "../src/adapters/web3signer";
 import { ContractReader } from "../src/adapters/subgraph";
 import { ContractWriter } from "../src/adapters/contract";
 import * as RouterFns from "../src/router";
+import { AuctionCache } from "../src/adapters/cache/auction";
+import { RouterCache } from "../src/adapters/cache";
 
+export let cacheMock: RouterCache;
 export let txServiceMock: SinonStubbedInstance<TransactionService>;
 export let messagingMock: SinonStubbedInstance<RouterNxtpNatsMessagingService>;
 export let contractReaderMock: ContractReader;
@@ -24,6 +28,9 @@ export let ctxMock: Context;
 export let isRouterContractMock: SinonStub<any, boolean>;
 export const routerAddress = routerAddrMock;
 export const signerAddress = mkAddress("0x123");
+export const chainAssetSwapPoolMapMock = new Map();
+chainAssetSwapPoolMapMock.set(1337, [mkAddress("0xc")]);
+chainAssetSwapPoolMapMock.set(1338, [mkAddress("0xf")]);
 
 export const mochaHooks = {
   async beforeEach() {
@@ -31,6 +38,11 @@ export const mochaHooks = {
     (walletMock as any).address = routerAddrMock; // need to do this differently bc the function doesnt exist on the interface
     walletMock.getAddress.resolves(routerAddrMock);
     walletMock.signMessage.resolves(sigMock);
+
+    let web3SignerWallet: SinonStubbedInstance<Web3Signer>;
+    web3SignerWallet = createStubInstance(Web3Signer);
+    (web3SignerWallet as any).address = routerAddrMock; // need to do this differently bc the function doesnt exist on the interface
+    web3SignerWallet.getAddress.resolves(routerAddrMock);
 
     txServiceMock = createStubInstance(TransactionService);
     txServiceMock.getBalance.resolves(parseEther("1"));
@@ -42,6 +54,7 @@ export const mochaHooks = {
     txServiceMock.calculateGasFeeInReceivingToken.resolves(BigNumber.from(100));
     txServiceMock.calculateGasFeeInReceivingTokenForFulfill.resolves(BigNumber.from(120));
     txServiceMock.getTokenPrice.resolves(BigNumber.from(1));
+    txServiceMock.getGasEstimate.resolves(BigNumber.from(24001));
 
     messagingMock = createStubInstance(RouterNxtpNatsMessagingService);
 
@@ -61,6 +74,13 @@ export const mochaHooks = {
           locked: BigNumber.from(0),
         },
       ]),
+    };
+
+    const auctionsCacheMock = createStubInstance(AuctionCache);
+    auctionsCacheMock.getOutstandingLiquidity.returns(BigNumber.from("0"));
+    auctionsCacheMock.addBid.resolves();
+    cacheMock = {
+      auctions: auctionsCacheMock as any,
     };
 
     contractWriterMock = {
@@ -86,6 +106,8 @@ export const mochaHooks = {
       isRouterContract: undefined,
       routerAddress,
       signerAddress,
+      cache: cacheMock,
+      chainAssetSwapPoolMap: chainAssetSwapPoolMapMock,
     };
 
     isRouterContractMock = stub(ctxMock, "isRouterContract").value(false);