+++ packages/router/test/lib/operations/status.spec.ts
@@ -7,13 +7,13 @@ import { ctxMock } from "../../globalTestHook";
 const { requestContext } = createLoggingContext("TEST", undefined, mkBytes32("0xabc"));
 
 describe("Status Operation", () => {
-  it("should work", async () => {
+  it("should work", () => {
     const ctxMockAssets = ctxMock.config.swapPools[0].assets;
-    const res = await getStatus(requestContext);
+    const res = getStatus(requestContext);
     const resSwapPools = res.swapPools;
 
     ctxMockAssets.forEach(({ assetId, chainId }) => {
-      expect(resSwapPools.get(chainId).includes(assetId)).to.be.true;
+      expect(resSwapPools[chainId].includes(assetId)).to.be.true;
     });
   });
 });