+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Next Release
 
+- [sdk] Return subgraph query in generic query method
+- [router] Track bids in memory to only bid 150% of available liquidity
+
+## 0.1.11
+
+- [contracts] Add ability to one-off verify contracts via hardhat
+- [router] Sanity check for router contracts existing on chain
+- [utils] Correctly determine the Gelato-supported chains
+- [router][txservice] Move sanity check of oracle address to txservice
+
 ## 0.1.10
 
 - [utils] Use the correct url for the subgraph health